// src/api/models/updatePastWeatherHistoriesRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Admin
 * Allta Admin 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */

export interface UpdatePastWeatherHistoriesRequest {
  /** 매장 ID */
  storeId: string;
  /** 지역코드 */
  stn: string;
  /** 시작일(YYYY-MM-DD) */
  startDate: string;
  /** 종료일(YYYY-MM-DD) */
  endDate: string;
}